requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.4
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.6.1
xgboost==2.1.4
lightgbm==4.6.0
catboost==1.2.7
tensorflow-cpu==2.17.1
joblib==1.4.2
python-dotenv==1.1.0
httpx==0.28.1
optuna==4.3.0
matplotlib==3.10.3
seaborn==0.13.2
plotly==6.1.2
